Fix OperationalMode Pilot
Thanks for the new version 4, great use og the AdminService. 
Pilot mode did not find any driverpackages using AdminService. Found out that the parameter for the name of the driverpackage is not 'PackageName' but 'Name'.
So this should fix Pilot mode, but also fix Production mode, so it's not using pilot drivers in production.